Move OSM tag reading code to its own method.
Since other data blocks than <way ...> can have a list of tags, it makes
sense that this code has its own method. OSM tags are all structured as
key/value pairs, so tags on, say, <node ...> are structured the same as
on a <way ...>, so the code is easily reusable.

This change is also in line with the Separation of Concerns principle.