Integrate UTMConverter into MapLoader
The MapLoader now converts the lat/lon data it reads into UTM
coordinates before constructing the nodes.

This commit also fixes the fact, that the map was unable to zoom when
it used lat/lon coordinates directly without converting them.

This commit also switches the minlat, maxlat, and minlon, maxlon of the
map bounds as they had been switched in commit bfcc180. This allowed the
map to be rendered when using lat/lon directly, but it was a hacky an
incorrect way to do it.