Allow data exported from OSM by different means
Selecting a box at the OSM website and clicking "Export" is only
available for smaller sections and formats data preceding the list of
nodes slightly differently, than when using the Overpass API (the node
list, way list, and relation list are identical, though) which allows
exporting larger sections of the map. This commit allows data exported
using both methods to be used by the application.