Remove redundant and unused code from RoadEdge
RoadEdge had its own instance of a roadname field in addition to the
one in its WayData. This was to avoid breaking a lot of tests in
NameSearchModelTest. RoadEdge has been altered and NameSearchModelTest
altered to accomodate the new RoadEdge implementation.

RoadEdge also had some leftover code from used in the original version
of the application. This code has been rendered redundant by the new
OSM data and the way it is stored by the application.