Assign proper RoadType to RoadEdges
Previously, just to have something working, every road was automatically
assigned the motorway type, so that it was easy to get an idea whether
road data was read correctly. This has been fixed to translate a highway
type from the OSM data to a RoadType enum object which is assigned to the
RoadEdge.