Properly read the direction of ways from data
WayData now has a oneway field instead of a direction field as this is
how OSM data handles way direction. This is also simpler (ie. to assume
that a oneway way is only traversible from start nodes to end nodes of
the constituent RoadEdges) than the old solution, which potentially allowed
a oneway RoadEdge to be traversed from end node to start node but not vice
versa which does not make sense.