Fix regression where road name search is broken
This regression was introduced in commit
cd2436f8d90efb19412a956a32fa2f7454613bfc and is a result of the fact
that Trie ordering was dependant on the toString() method of the ordered
object type. The user would have to search the toString() representation
of the road, so when that representation changed from just <name-of-road> to
[RoadEdge roadname="<name-of-road>"..., the user would have to write
[RoadEdge roadname="<name-of-road>" everytime a search was performed for
the Trie to find anything.

With this commit, Trie ordering now uses a separate string representation
as this allows for a more informative toString() representation while
allowing the user to search simply by the name of the road.