Fix route search
LiveComboBox contained Strings instead of RoadEdges, so the search
would always fail as no RoadEdges was found to find a path between.

LiveComboBox was also dependent on the toString() of the
objects it contained, as it used it to determine which string
representation to show in the ComboBox. Now, LiveComboBox no longer
uses toString(), but the Trie representation which is also the
representation that the user entered query string is matched against.
At this point, the Trie representation is only the name of the RoadEdge.