Fix LiveComboBox return key breaking route search
When using the return key to select a highlighted road in the dropdown,
route search would complain that no end points for the search had been
set, while it would work fine when using the mouse to click a
highlighted road.
The return key will now properly select a highlighted road.